part3_auto_apply_agent/requirements.txt
openai>=1.12.0
playwright>=1.40.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
aiohttp>=3.9.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

